Remove old EventCodes

+++ StatisticsAnalysisTool/StatisticsAnalysisTool/Network/Handler/PartySilverGainedEventHandler.cs
@@ -1,4 +1,5 @@
 ﻿using Albion.Network;
+using StatisticsAnalysisTool.Enumerations;
 using StatisticsAnalysisTool.Network.Events;
 using System.Diagnostics;
 using System.Threading.Tasks;
@@ -7,7 +8,7 @@ namespace StatisticsAnalysisTool.Network.Handler
 {
     public class PartySilverGainedEventHandler : EventPacketHandler<PartySilverGainedEvent>
     {
-        public PartySilverGainedEventHandler() : base(EventCodes.PartySilverGained) { }
+        public PartySilverGainedEventHandler() : base((int) EventCodes.PartySilverGained) { }
 
         protected override async Task OnActionAsync(PartySilverGainedEvent value)
         {
