+++ StatisticsAnalysisTool/StatisticsAnalysisTool/Network/Handler/TakeSilverEventHandler.cs
@@ -1,11 +1,12 @@
 ﻿using Albion.Network;
+using StatisticsAnalysisTool.Enumerations;
 using System.Threading.Tasks;
 
 namespace StatisticsAnalysisTool.Network.Handler
 {
     public class TakeSilverEventHandler : EventPacketHandler<TakeSilverEvent>
     {
-        public TakeSilverEventHandler() : base(EventCodes.TakeSilver) { }
+        public TakeSilverEventHandler() : base((int) EventCodes.TakeSilver) { }
 
         protected override async Task OnActionAsync(TakeSilverEvent value)
         {
