+++ StatisticsAnalysisTool/StatisticsAnalysisTool/Network/Handler/UpdateMoneyEventHandler.cs
@@ -10,7 +10,7 @@ namespace StatisticsAnalysisTool.Network.Handler
         private readonly TrackingController _trackingController;
         private readonly SilverCountUpTimer _silverCountUpTimer;
 
-        public UpdateMoneyEventHandler(TrackingController trackingController, SilverCountUpTimer silverCountUpTimer) : base(EventCodes.UpdateMoney)
+        public UpdateMoneyEventHandler(TrackingController trackingController, SilverCountUpTimer silverCountUpTimer) : base((int) EventCodes.UpdateMoney)
         {
             _trackingController = trackingController;
             _silverCountUpTimer = silverCountUpTimer;
